Refactor shell completion from global option to dedicated command
- Replace --install-completion global option with `rxiv completion` command
- Add new completion.py module with install_shell_completion functionality
- Update CLI help documentation to showcase new completion command
- Update all tests to use new completion command syntax
- Enable completion installation for bash, zsh, and fish shells
- Maintain backward compatibility while improving CLI structure
