--- .gitmodules
@@ -1,9 +1,0 @@
-[submodule "submodules/homebrew-rxiv-maker"]
-	path = submodules/homebrew-rxiv-maker
-	url = https://github.com/henriqueslab/homebrew-rxiv-maker.git
-[submodule "submodules/scoop-rxiv-maker"]
-	path = submodules/scoop-rxiv-maker
-	url = https://github.com/henriqueslab/scoop-rxiv-maker.git
-[submodule "submodules/vscode-rxiv-maker"]
-	path = submodules/vscode-rxiv-maker
-	url = https://github.com/HenriquesLab/vscode-rxiv-maker.git
