+++ EXAMPLE_MANUSCRIPT/FIGURES/SFigure__architecture.mmd
@@ -55,7 +55,7 @@ flowchart TD
         D3["📋 Build Artifacts<br>• Auxiliary files<br>• Build logs<br>• Debug information"]
   end
 
- subgraph DEPLOY["🚀 Deployment Environments"]
+ subgraph DEPLOY["🚀 Deployment Env."]
     direction TB
     subgraph LOCAL["💻 Local Execution"]
       direction TB
