+++ pyproject.toml
@@ -36,8 +36,6 @@ dependencies = [
     "PyYAML>=6.0.0",
     "python-dotenv>=1.0.0",
     "crossref-commons>=0.0.7",
-    "lazydocs>=0.4.8",
-    "folder2md4llms>=0.4.28",
     "click>=8.0.0",
     "rich>=13.0.0",
     "rich-click>=1.6.0",
@@ -66,7 +64,6 @@ dev = [
     "coverage[toml]>=7.0",
     "pytest-notebook>=0.10.0",
     "nbstripout>=0.7.1",
-    "lazydocs>=0.4.8",
     "pytest-timeout>=2.0.0",
     "pytest-mock>=3.10.0",
     "pytest-sugar>=0.9.7",
@@ -100,7 +97,6 @@ dev = [
     "pytest-notebook>=0.10.0",
     "nbstripout>=0.7.1",
     "pre-commit>=4.0.0",
-    "lazydocs>=0.4.8",
     "hatch>=1.14.1",
     "mkdocs>=1.4.0",
     "mkdocs-material>=9.0.0",
