+++ src/rxiv_maker/__version__.py
@@ -1,7 +1,7 @@
 """Version information for rxiv-maker."""
 
-__version__ = "1.4.17"
-__version_tuple__ = (1, 4, 17)
+__version__ = "1.4.19"
+__version_tuple__ = (1, 4, 19)
 
 # Note: Docker images v1.8+ use mermaid.ink API for diagram processing
 # This improves cross-platform compatibility and performance
