+++ src/rxiv_maker/cli/commands/__init__.py
@@ -5,6 +5,7 @@ from .bibliography import bibliography
 from .build import build as pdf
 from .check_installation import check_installation
 from .clean import clean
+from .completion import completion_cmd
 from .figures import figures
 from .init import init
 from .install_deps import install_deps
@@ -19,6 +20,7 @@ __all__ = [
     "pdf",
     "check_installation",
     "clean",
+    "completion_cmd",
     "figures",
     "init",
     "install_deps",
