+++ src/rxiv_maker/utils/platform.py
@@ -182,6 +182,7 @@ class PlatformDetector:
             return "conda-forge" in result.stdout
         except (subprocess.TimeoutExpired, FileNotFoundError, Exception):
             return False
+
     def run_command(self, cmd: str, shell: bool = True, **kwargs) -> subprocess.CompletedProcess:
         """Run a command with platform-appropriate settings."""
         if self.is_windows():
